General improvement for Android and database management
Now we can upload a picture for each beer from the android app.
Improve the search for a new brewery (don't care about the case anymore).
Now we know which user created a brewery.
Create the database maangement
Improve the search for tjhe beers (don't care about the case, and fix the path for the label)
Add the routes for the database management.
In the beer model, add the fallback cases in case one of the field is not filled.
From the brewery model we now have more easy method to get the beers associated and the number of beers.
Add the create in the brewery model.
Add the role in the users model (so people can have the right to edit databasse)
Create the UserRole model.